requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-appconf==1.0.6
django-auditlog==3.0.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-filter==24.3
django-tables2==2.7.5
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
rcssmin==1.1.2
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
